integrate cffi module and original C module
